+++ trainers-workload-service/src/main/java/com/trainersworkloadservice/repository/YearRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface YearRepository extends JpaRepository<YearEntity, Long> {
-    Optional<YearEntity> findByTrainerIdAndWorkYear(Long trainerId, short workYear);
+    Optional<YearEntity> findByTrainerIdAndWorkYear(Long trainerId, int workYear);
 
-    boolean existsByTrainerIdAndWorkYear(Long trainerId, short workYear);
+    boolean existsByTrainerIdAndWorkYear(Long trainerId, int workYear);
 }
